+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Don't display alt text if custom emoji doesn't load.
+- Use "Conversation" visibility by default when replying in followers-only conversation.
 
 ## [3.11.0] - 2024-12-05